Changed dashboard.index

+++ database/migrations/2019_10_14_115727_create_seismic_events_table.php
@@ -20,7 +20,7 @@ class CreateSeismicEventsTable extends Migration
             $table->double("long");
             $table->double("lat");
 
-            $table->integer("mag");
+            $table->double("mag");
             $table->string("type");
             $table->date("date");
             $table->string("place");
